Fix OAuth constructor to correctly allow request token creation
Smallest change to get the system to work, so far done extensive testing and have not found any drawbacks but I imagine there is a 'better' way which would require a bit more changes within SugarOAuthServer. setRequestTokenPath would set the requestPath but when needsToken() was called it checked is_request which was never set (and so told the system that a token was indeed required). Function I switched to ( isRequestTokenEndpoint ) sets the is_request variable and therefore needsToken() correctly responds false.

needsToken itself behaves oddly... it returns false if is_request is not empty or true if requestPath is empty. However when you SET requestPath (like the previous implementation did) it did neither and instead returned a url (in which case any thing which looked for (if $this->needsToken() would get a 'not false' back and expect a token despite the request being for a request token... That is likely not how it should be written but would be upstream with Zend?